Move ConciergeClient.convert_ms_object to utils module
For 2 reasons.

1.  I've got a use case where I want to use convert_ms_object() but I
don't have a ConciergeClient.  So make it a class method?  Well, I
thought of that, but ...

2.  convert_ms_object() doesn't do anything with its first argument
whether it's `self` or `cls` so it can really stand alone, separate
from ConciergeClient.  Hence new home in utils.
